requirement.txt
python-dotenv
pandas 
scikit-learn
streamlit
streamlit-authenticator
openpyxl
xlsxwriter
pyspellchecker
spacy
unidecode